pyproject.toml
[tool.poetry]
name = "key-switcheroo"
version = "0.1.0"
description = "Rotate SSH keys, stored in the cloud!"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "switcheroo"}]

[tool.poetry.dependencies]
python = "^3.11"
paramiko = "^3.2.0"
pycryptodome = "^3.18.0"
psutil = "^5.9.5"
boto3 = "^1.27.0"

[tool.poetry.group.dev.dependencies]
PyHamcrest = "^2.0.4"
pyright = "^1.1.316"
pylint = "^2.17.4"
black = "^23.3.0"
boto3-stubs = {version = "1.27.0", extras = ["s3", "sts"]}

[tool.poetry.scripts]

switcheroo_publish = "switcheroo.ssh.scripts.publish:main"
switcheroo_retrieve = "switcheroo.ssh.scripts.retrieve:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
